+++ core/src/cats/sql_list.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2000-2009 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -227,12 +227,10 @@ void BareosDb::ListMediaRecords(JobControlRecord* jcr,
 
   EscapeString(jcr, esc, mdbr->VolumeName, strlen(mdbr->VolumeName));
 
-  /*
-   * There is one case where ListMediaRecords() is called from SelectMediaDbr()
+  /* There is one case where ListMediaRecords() is called from SelectMediaDbr()
    * with the range argument set to NULL. To avoid problems, we set the range to
    * an empty string if range is set to NULL. Otherwise it would result in
-   * malformed SQL queries.
-   */
+   * malformed SQL queries. */
   if (range == NULL) { range = ""; }
 
   if (count) {
@@ -436,12 +434,11 @@ void BareosDb::ListLogRecords(JobControlRecord* jcr,
   }
 
   if (type != VERT_LIST) {
-    /*
-     * When something else then a vertical list is requested set the list type
+    /* When something else then a vertical list is requested set the list type
      * to RAW_LIST e.g. non formated raw data as that makes the only sense for
      * the logtext output. The logtext already has things like \n etc in it
-     * so we should just dump the raw content out for the best visible output.
-     */
+     * so we should just dump the raw content out for the best visible
+     * output. */
     type = RAW_LIST;
   }
 
@@ -473,12 +470,11 @@ void BareosDb::ListJoblogRecords(JobControlRecord* jcr,
   } else {
     FillQuery(SQL_QUERY::list_joblog_2, edit_int64(JobId, ed1), range);
     if (type != VERT_LIST) {
-      /*
-       * When something else then a vertical list is requested set the list type
+      /* When something else then a vertical list is requested set the list type
        * to RAW_LIST e.g. non formated raw data as that makes the only sense for
        * the logtext output. The logtext already has things like \n etc in it
-       * so we should just dump the raw content out for the best visible output.
-       */
+       * so we should just dump the raw content out for the best visible
+       * output. */
       type = RAW_LIST;
     }
   }
@@ -525,7 +521,7 @@ void BareosDb::ListJobRecords(JobControlRecord* jcr,
                               const char* range,
                               const char* clientname,
                               std::vector<char> jobstatuslist,
-                              int joblevel,
+                              std::vector<char> joblevels,
                               std::vector<char> jobtypes,
                               const char* volumename,
                               const char* poolname,
@@ -563,8 +559,9 @@ void BareosDb::ListJobRecords(JobControlRecord* jcr,
     PmStrcat(selection, temp.c_str());
   }
 
-  if (joblevel) {
-    temp.bsprintf("AND Job.Level = '%c' ", joblevel);
+  if (!joblevels.empty()) {
+    std::string levels = CreateDelimitedStringForSqlQueries(joblevels, ',');
+    temp.bsprintf("AND Job.Level in (%s) ", levels.c_str());
     PmStrcat(selection, temp.c_str());
   }