+++ core/src/dird/ua_select.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2001-2012 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -741,11 +741,9 @@ bool SelectPoolDbr(UaContext* ua, PoolDbRecord* pr, const char* argk)
 
   new (&opr) PoolDbRecord();  // placement new instead of memset
 
-  /*
-   * *None* is only returned when selecting a recyclepool, and in that case
+  /* *None* is only returned when selecting a recyclepool, and in that case
    * the calling code is only interested in opr.Name, so then we can leave
-   * pr as all zero.
-   */
+   * pr as all zero. */
   if (!bstrcmp(name, _("*None*"))) {
     bstrncpy(opr.Name, name, sizeof(opr.Name));
 
@@ -841,11 +839,9 @@ bool SelectStorageDbr(UaContext* ua, StorageDbRecord* sr, const char* argk)
 
   new (&osr) StorageDbRecord();  // placement new instead of memset
 
-  /*
-   * *None* is only returned when selecting a recyclestorage, and in that case
+  /* *None* is only returned when selecting a recyclestorage, and in that case
    * the calling code is only interested in osr.Name, so then we can leave
-   * sr as all zero.
-   */
+   * sr as all zero. */
   if (!bstrcmp(name, _("*None*"))) {
     bstrncpy(osr.Name, name, sizeof(osr.Name));
 
@@ -990,9 +986,8 @@ PoolResource* get_pool_resource(UaContext* ua)
 // List all jobs and ask user to select one
 int SelectJobDbr(UaContext* ua, JobDbRecord* jr)
 {
-  ua->db->ListJobRecords(ua->jcr, jr, "", NULL, std::vector<char>{}, 0,
-                         std::vector<char>{}, NULL, NULL, 0, 0, 0, ua->send,
-                         HORZ_LIST);
+  ua->db->ListJobRecords(ua->jcr, jr, "", NULL, {}, {}, {}, nullptr, nullptr, 0,
+                         0, 0, ua->send, HORZ_LIST);
   if (!GetPint(ua, _("Enter the JobId to select: "))) { return 0; }
 
   jr->JobId = ua->int64_val;
@@ -1610,10 +1605,8 @@ alist<JobId_t*>* select_jobs(UaContext* ua, const char* reason)
     }
   }
 
-  /*
-   * If we didn't select any Jobs using jobid, job or ujobid keywords try
-   * other selections.
-   */
+  /* If we didn't select any Jobs using jobid, job or ujobid keywords try
+   * other selections. */
   if (cnt == 0) {
     char buf[1000];
     int tjobs = 0; /* Total # number jobs */
@@ -1673,10 +1666,8 @@ alist<JobId_t*>* select_jobs(UaContext* ua, const char* reason)
         }
       }
 
-      /*
-       * Select from all available Jobs the Jobs matching the selection
-       * criterium.
-       */
+      /* Select from all available Jobs the Jobs matching the selection
+       * criterium. */
       foreach_jcr (jcr) {
         if (jcr->JobId == 0) { /* This is us */
           continue;
@@ -1719,10 +1710,8 @@ alist<JobId_t*>* select_jobs(UaContext* ua, const char* reason)
         goto bail_out;
       }
 
-      /*
-       * Only ask for confirmation when not in batch mode and there is no yes
-       * on the cmdline.
-       */
+      /* Only ask for confirmation when not in batch mode and there is no yes
+       * on the cmdline. */
       if (!ua->batch && FindArg(ua, NT_("yes")) == -1) {
         if (!GetYesno(ua, _("Confirm cancel (yes/no): ")) || !ua->pint32_val) {
           goto bail_out;
@@ -1808,10 +1797,8 @@ bool GetUserSlotList(UaContext* ua,
   bstrncpy(search_argument, argument, sizeof(search_argument));
   i = FindArgWithValue(ua, search_argument);
   if (i == -1) { /* not found */
-    /*
-     * See if the last letter of search_argument is a 's'
-     * When it is strip it and try if that argument is given.
-     */
+    /* See if the last letter of search_argument is a 's'
+     * When it is strip it and try if that argument is given. */
     len = strlen(search_argument);
     if (len > 0 && search_argument[len - 1] == 's') {
       search_argument[len - 1] = '\0';
@@ -1990,17 +1977,22 @@ bool GetUserJobStatusSelection(UaContext* ua, std::vector<char>& jobstatuslist)
   return true;
 }
 
-bool GetUserJobLevelSelection(UaContext* ua, int* joblevel)
+bool GetUserJobLevelSelection(UaContext* ua, std::vector<char>& joblevel_list)
 {
   int i;
 
-  if ((i = FindArgWithValue(ua, NT_("joblevel"))) >= 0) {
-    if (strlen(ua->argv[i]) == 1 && ua->argv[i][0] >= 'A'
-        && ua->argv[i][0] <= 'z') {
-      *joblevel = ua->argv[i][0];
-    } else {
-      /* invalid joblevel */
-      return false;
+  if (((i = FindArgWithValue(ua, NT_("joblevel"))) >= 0)
+      || ((i = FindArgWithValue(ua, NT_("level"))) >= 0)) {
+    std::vector<std::string> joblevelinput_list
+        = split_string(ua->argv[i], ',');
+
+    for (const auto& level : joblevelinput_list) {
+      if (level.size() == 1 && level[0] >= 'A' && level[0] <= 'z') {
+        joblevel_list.push_back(level[0]);
+      } else {
+        /* invalid joblevel */
+        return false;
+      }
     }
   }
   return true;